Continued work on CSV_importer
Working on implementing the many folders and/or many files logic and data handling features.

Intended features will allow the function to load data from any number of folders and/or any number of files and put them together based on options passed into the function.

Currently completed the following cases:
1 Folder 1 File
n Folders 1 File
1 Folder n Files

Need to work on:
n Folders n Files

This is challenging because the Folders and Files may be concatenated or merged in different orders, and it must be handeled such that the dataframe1 and dataframe2 in the pd.concat or pd.merge method are incompatible sizes or shapes.