Fixed missing file issue
modified _load_csv to return an empty df if the file is not found. This will allow the _combine functions to operate normally, and when all is said and done the appropriate missing data is turned into NaN because pandas is amazing and appears to be resizing the main df if we add a big one to a small one.